DRILL-6588: Make Sys tables o nullable datatypes
This is to address the generic problem of NULL values being projected as a string because of all datatypes being non-nullable.
Although there would be columns that should be non-nullable (e.g. sys.options.name), System tables are immutable, so a NULL value for those columns would never appear. Hence, this patch only applies to System tables.