DRILL-6960: AutoLimit the size of ResultSet for a WebUI (or REST) client
Provide an option for browser-based (actually, any REST API) query submissions to limit the number of records being fetched from the server.
The WebUI will render a warning indicating that the result set size was artificially limited 
Additional changes include all warnings for empty or incorrect fields not being displayed as an alert but as an error message.